Exception Handling

+++ src/main/java/com/projetos/dscommerce/servicies/ProductService.java
@@ -3,16 +3,15 @@ package com.projetos.dscommerce.servicies;
 import com.projetos.dscommerce.dto.ProductDTO;
 import com.projetos.dscommerce.entities.Product;
 import com.projetos.dscommerce.repositories.ProductRepository;
+import com.projetos.dscommerce.servicies.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import javax.swing.text.html.Option;
-import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -22,9 +21,9 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Product product =  repository.findById(id).get();
+        Product product =  repository.findById(id).
+                orElseThrow(() -> new ResourceNotFoundException("Resource not Found with id " + id));
         return new ProductDTO(product);
-
     }
 
     @Transactional(readOnly = true)
@@ -44,15 +43,30 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }
+        catch (EntityNotFoundException e ){
+            throw  new ResourceNotFoundException("Resource not found");
+        }
 
     }
 
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if (!repository.existsById(id)){
+            throw new ResourceNotFoundException("Resource not Found");
+        }
+        try {
+            repository.deleteById(id);
+
+        }
+        catch (DataIntegrityViolationException e){
+            throw new DataIntegrityViolationException("Data Integrity Violation Excpetion");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
